Create README.md
Java FXML Project
The Atomic Appointment Manager “Power Tools for Successful Professionals”
Coming very soon (expected release time 2/17/2020), with the upgraded focus view that keeps your edge sharp with your appointments sorted by type, and quantity both weekly and monthly. This feature helps you focus your productivity towards appointments that generate wholeness, prosperity, and balance in your life (forget about chasing numbers at the end of the quarter, or the end of the month with your wise choice with this dynamic tool).
This is attached to a MySQL database with real-time reflectivity in both customer and appointment manager screens.
Both HashMap’s and Lambdas are used, and this application incorporates both English and French (might have been influenced by Valentine’s day) depending upon the location the user logins in at.

Some of the amazing features include:
Schedule for each consultant.
The number of appointments by type with both week and month view.
An appointment can only be scheduled inside your business hours.
No overlapping appointments are allowed to be booked.
Ability to add, update and delete appointments, whilst capturing the type and a link to a customer record in the database.
Recording timestamps for user logins in a .txt file.
Providing an Alert if an appointment is scheduled within 15 minutes of the user's login.